:sparkles: (Koin) : Changed Koin modules to new pattern.
