+++ app/src/main/java/com/kidemma/common/utils/Utils.kt
@@ -1,0 +1,32 @@
+package com.kidemma.common.utils
+
+import com.google.gson.GsonBuilder
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import java.util.concurrent.TimeUnit
+
+inline fun <reified T> createWebService(
+    okHttpClient: OkHttpClient,
+    baseUrl: String
+): T {
+    val retrofit = Retrofit.Builder()
+        .baseUrl(baseUrl)
+        .addConverterFactory(GsonConverterFactory.create(GsonBuilder().create()))
+        .client(okHttpClient)
+        .build()
+    return retrofit.create(T::class.java)
+}
+
+fun createHttpClient(): OkHttpClient {
+    val interceptor = HttpLoggingInterceptor().apply {
+        level = HttpLoggingInterceptor.Level.BODY
+    }
+
+    val builder = OkHttpClient.Builder()
+        .readTimeout(60, TimeUnit.SECONDS)
+        .addInterceptor(interceptor)
+
+    return builder.build()
+}
