+++ gradle/libs.versions.toml
@@ -14,7 +14,6 @@ koin = "4.1.1"
 loggingInterceptor = "5.3.2"
 retrofit = "3.0.0"
 coroutines = "1.10.2"
-rxjava2 = "2.9.0"
 lottie = "6.7.1"
 glide = "5.0.5"
 navigationCompose = "2.9.7"
